requirements.txt
google-cloud-bigquery
mysql-connector-python
pandas